requirements.txt
faiss-cpu
numpy
sentence-transformers
google-generativeai
uvicorn